requirements.txt
autopep8==1.4.4
cffi==1.13.2
cryptography==2.8
Django==2.2.7
django-annoying==0.10.5
django-rest-knox==4.1.0
djangorestframework==3.10.3
gunicorn==20.0.0
pep8==1.7.1
Pillow==6.2.1
psycopg2==2.8.4
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pycparser==2.19
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
